Initial commit - Library code & use examples
This is a library project that I'm working on, mainly as a personal
training task but also for later use in other projects of mine.
The main features at the moment are:
* UDP multithreaded server (IPv4 + IPv6)
* UDP client
* Generic A* algorithm (shipped as a class template)
* Logging facility
* Exceptions
* Random numbers generation
* 2D points manipulation

The code compiles & runs on Linux & Windows platforms (at least those I've
tested, ~= Linux 4.x flavors with g++4.9 & Windows 7 with Visual C++).

Some code examples for the AStar, UDPServer & UDPClient modules can be
found in the examples/ folder but they still lack some documentation (work
in progress, I promise).
Please note that the AStar example requires SFML headers and library to
compile, as it features a nicely undocumented GUI to test the path finder.
:)